refactor modules, list user's joined groups in GroupScreen
