+++ src/components/group/GroupScreen.js
@@ -3,18 +3,17 @@ import {
 	StyleSheet,
 	Text,
 	View,
-  ScrollView,
-  TouchableOpacity
+  ScrollView
 } from 'react-native';
-import PropTypes from 'prop-types';
-import { Button } from 'react-native-elements';
 import { observer } from 'mobx-react';
 
 import AppHeaderSwitch from '../../custom/AppHeaderSwitch';
 import BottomBar from '../BottomBar';
-import { container as containerStyles, iconButton, subSection } from '../../styles';
-import { colors, button, text } from '../../styles/theme';
+import { container as containerStyles, subSection } from '../../styles';
+import { colors, text } from '../../styles/theme';
 import GroupStore from '../../stores/group';
+import UserStore from '../../stores/user';
+import Group from './Group';
 
 @observer
 export class GroupScreen extends React.Component {
@@ -30,12 +29,31 @@ export class GroupScreen extends React.Component {
 
 	render() {
     const groups = GroupStore.groups;
+    const userGroups = groups.filter(group => {
+      return UserStore.joinedGroups.filter(g => g.id === group.id).length > 0;
+    })
 
     return (
       <View style={styles.container}>
         <AppHeaderSwitch viewName="Groups" />
 
         <ScrollView style={styles.contentContainer}>
+
+          <View>
+            <View style={styles.subHeader}>
+              <Text style={styles.subHeaderText}>Your groups</Text>
+            </View>
+            <View>
+              {userGroups.map(group => (
+                <Group
+                  key={group.id}
+                  group={group}
+                  navigate={this.getNavigate(group)}
+                />
+              ))}
+            </View>
+          </View>
+
           <View>
             <View style={styles.subHeader}>
               <Text style={styles.subHeaderText}>Common groups</Text>
@@ -50,6 +68,7 @@ export class GroupScreen extends React.Component {
               ))}
             </View>
           </View>
+
         </ScrollView>
 
         <BottomBar currentView='Groups' />
@@ -57,32 +76,6 @@ export class GroupScreen extends React.Component {
     );
   }
 }
-
-const Group = (props) => {
-  const { group, navigate  } = props;
-
-  return (
-    <TouchableOpacity onPress={navigate}>
-      <View style={styles.group}>
-        <Text style={styles.groupText}>{group.name}</Text>
-        <Button
-          containerViewStyle={{marginRight: 0}}
-          iconRight={{
-            name: 'chevron-right',
-            size: 25,
-            style: {color: colors.black}
-          }}
-          buttonStyle={styles.groupButton}
-        />
-      </View>
-    </TouchableOpacity>
-  );
-}
-
-Group.propTypes = {
-  navigate: PropTypes.func,
-  group: PropTypes.object
-};
 
 const styles = StyleSheet.create({
   container: {
@@ -98,22 +91,6 @@ const styles = StyleSheet.create({
   subHeaderText: {
     fontSize: 18,
     fontWeight: text.boldWeight
-  },
-  group: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    padding: 16,
-    borderBottomWidth: 0.5,
-    borderColor: colors.line
-  },
-  groupText: {
-    fontSize: 16
-  },
-  groupButton: {
-    padding: 0,
-    backgroundColor: 'transparent',
-    ...iconButton
   }
 });
